moved fee view to separate file tweaked margings
